remove repositories from mappers
